Remove profile action to profile controller
