+++ src/test/java/ru/javaluxurywatches/controller/UserControllerTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import ru.javaluxurywatches.controller.user.UserController;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
